CL: Limit max work size/offset to UINT32_MAX
Set a hard limit of UINT32_MAX on work item/offset for OpenCL's
clEnqueueNDRangeKernel cmd for all backends (simplifies handling).

Bug: angleproject:42267067
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6191436
Commit-Queue: Austin Annestrand <a.annestrand@samsung.com>

+++ src/libANGLE/renderer/vulkan/CLKernelVk.cpp
@@ -292,9 +292,9 @@ angle::Result CLKernelVk::getOrCreateComputePipeline(vk::PipelineCacheAccess *pi
     ASSERT(workgroupSize[0] != 0);
     ASSERT(workgroupSize[1] != 0);
     ASSERT(workgroupSize[2] != 0);
-    (*workgroupCountOut)[0] = static_cast<uint32_t>((ndrange.globalWorkSize[0] / workgroupSize[0]));
-    (*workgroupCountOut)[1] = static_cast<uint32_t>((ndrange.globalWorkSize[1] / workgroupSize[1]));
-    (*workgroupCountOut)[2] = static_cast<uint32_t>((ndrange.globalWorkSize[2] / workgroupSize[2]));
+    (*workgroupCountOut)[0] = (ndrange.globalWorkSize[0] / workgroupSize[0]);
+    (*workgroupCountOut)[1] = (ndrange.globalWorkSize[1] / workgroupSize[1]);
+    (*workgroupCountOut)[2] = (ndrange.globalWorkSize[2] / workgroupSize[2]);
 
     // Populate program specialization constants (if any)
     uint32_t constantDataOffset = 0;
@@ -308,22 +308,22 @@ angle::Result CLKernelVk::getOrCreateComputePipeline(vk::PipelineCacheAccess *pi
                 specConstantData.push_back(ndrange.workDimensions);
                 break;
             case SpecConstantType::WorkgroupSizeX:
-                specConstantData.push_back(static_cast<uint32_t>(workgroupSize[0]));
+                specConstantData.push_back(workgroupSize[0]);
                 break;
             case SpecConstantType::WorkgroupSizeY:
-                specConstantData.push_back(static_cast<uint32_t>(workgroupSize[1]));
+                specConstantData.push_back(workgroupSize[1]);
                 break;
             case SpecConstantType::WorkgroupSizeZ:
-                specConstantData.push_back(static_cast<uint32_t>(workgroupSize[2]));
+                specConstantData.push_back(workgroupSize[2]);
                 break;
             case SpecConstantType::GlobalOffsetX:
-                specConstantData.push_back(static_cast<uint32_t>(ndrange.globalWorkOffset[0]));
+                specConstantData.push_back(ndrange.globalWorkOffset[0]);
                 break;
             case SpecConstantType::GlobalOffsetY:
-                specConstantData.push_back(static_cast<uint32_t>(ndrange.globalWorkOffset[1]));
+                specConstantData.push_back(ndrange.globalWorkOffset[1]);
                 break;
             case SpecConstantType::GlobalOffsetZ:
-                specConstantData.push_back(static_cast<uint32_t>(ndrange.globalWorkOffset[2]));
+                specConstantData.push_back(ndrange.globalWorkOffset[2]);
                 break;
             default:
                 UNIMPLEMENTED();
